fix(chat): evict derived transcript blocks
