+++ src/renderer/src/app/ai/transcriptWindowStore.ts
@@ -19,6 +19,11 @@ interface SessionWindow {
     readonly touchedAt: Map<string, number>;
 }
 
+export interface EvictedTranscriptBlock {
+    readonly blockId: string;
+    readonly sessionId: string;
+}
+
 export interface TranscriptWindowSnapshot {
     readonly blocksById: ReadonlyMap<string, AiTranscriptBlock>;
     readonly generation: number;
@@ -28,6 +33,7 @@ export interface TranscriptWindowSnapshot {
 
 export class TranscriptWindowStore {
     private readonly sessions = new Map<string, SessionWindow>();
+    private readonly evictedBlocks: EvictedTranscriptBlock[] = [];
     private readonly evictedSessionIds = new Set<string>();
 
     constructor(
@@ -49,6 +55,7 @@ export class TranscriptWindowStore {
                 session.blocks.delete(blockId);
                 session.touchedAt.delete(blockId);
                 session.protectedBlockIds.delete(blockId);
+                this.recordEviction(sessionId, blockId);
             }
         }
         session.metadata = metadata;
@@ -64,12 +71,17 @@ export class TranscriptWindowStore {
     }
 
     clear(sessionId: string): void {
+        const session = this.sessions.get(sessionId);
+        for (const blockId of session?.blocks.keys() ?? []) {
+            this.recordEviction(sessionId, blockId);
+        }
         this.sessions.delete(sessionId);
         this.evictedSessionIds.delete(sessionId);
     }
 
     reset(): void {
         this.sessions.clear();
+        this.evictedBlocks.length = 0;
         this.evictedSessionIds.clear();
     }
 
@@ -121,6 +133,12 @@ export class TranscriptWindowStore {
         return sessionIds;
     }
 
+    takeEvictedBlocks(): readonly EvictedTranscriptBlock[] {
+        const blocks = [...this.evictedBlocks];
+        this.evictedBlocks.length = 0;
+        return blocks;
+    }
+
     private evict(): void {
         while (this.residentEntryCount() > this.maxResidentEntries) {
             const candidate = this.findEvictionCandidate(false) ??
@@ -128,7 +146,7 @@ export class TranscriptWindowStore {
             if (!candidate) return;
             candidate.session.blocks.delete(candidate.blockId);
             candidate.session.touchedAt.delete(candidate.blockId);
-            this.evictedSessionIds.add(candidate.sessionId);
+            this.recordEviction(candidate.sessionId, candidate.blockId);
             incrementChatPerformanceCounter("transcript_blocks_evicted");
         }
     }
@@ -167,6 +185,11 @@ export class TranscriptWindowStore {
             count += residentEntryCount(session);
         }
         return count;
+    }
+
+    private recordEviction(sessionId: string, blockId: string): void {
+        this.evictedSessionIds.add(sessionId);
+        this.evictedBlocks.push({ blockId, sessionId });
     }
 
     private sessionFor(sessionId: string): SessionWindow {
